+++ packages/contracts/contracts/registry/DAORegistry.sol
@@ -7,9 +7,9 @@ import "./InterfaceBasedRegistry.sol";
 
 /// @title Register your unique DAO name
 /// @author Aragon Association - 2021
-/// @notice This contract provides the possiblity to register a DAO by a unique name.
+/// @notice This contract provides the possiblity to register a DAO.
 contract DAORegistry is InterfaceBasedRegistry {
-    bytes32 public constant REGISTER_ROLE = keccak256("REGISTER_ROLE");
+    bytes32 public constant REGISTER_DAO_ROLE = keccak256("REGISTER_DAO_ROLE");
 
     /// @notice Emitted when a new DAO is registered
     /// @param dao The address of the DAO contract
@@ -28,11 +28,12 @@ contract DAORegistry is InterfaceBasedRegistry {
     /// @dev A name is unique within the Aragon DAO framework and can get stored here
     /// @param name The name of the DAO
     /// @param dao The address of the DAO contract
+    /// @param creator The address of the creator
     function register(
         string calldata name,
         IDAO dao,
         address creator
-    ) external auth(REGISTER_ROLE) {
+    ) external auth(REGISTER_DAO_ROLE) {
         _register(address(dao));
 
         emit NewDAORegistered(dao, creator, name);